move chain state transition code in separate directory
