+++ packages/lodestar/src/chain/blocks/stateTransition/events.ts
@@ -1,0 +1,55 @@
+import {ChainEvent, ChainEventEmitter, IBlockJob} from "../..";
+import {byteArrayEquals} from "@chainsafe/ssz";
+import {IBlockSummary, IForkChoice} from "@chainsafe/lodestar-fork-choice";
+import {computeEpochAtSlot, ZERO_HASH} from "@chainsafe/lodestar-beacon-state-transition";
+import {ITreeStateContext} from "../../../db/api/beacon/stateContextCache";
+import {assert} from "@chainsafe/lodestar-utils";
+
+/**
+ * Emits a properly formed "checkpoint" event, given a checkpoint state context
+ *
+ * This will throw an error if the checkpoint state is not a valid checkpoint state, eg: NOT the first slot in an epoch.
+ */
+export function emitCheckpointEvent(emitter: ChainEventEmitter, checkpointStateContext: ITreeStateContext): void {
+  const config = checkpointStateContext.epochCtx.config;
+  const slot = checkpointStateContext.state.slot;
+  assert.true(slot % config.params.SLOTS_PER_EPOCH === 0, "Checkpoint state slot must be first in an epoch");
+  const blockHeader = config.types.BeaconBlockHeader.clone(checkpointStateContext.state.latestBlockHeader);
+  if (config.types.Root.equals(blockHeader.stateRoot, ZERO_HASH)) {
+    blockHeader.stateRoot = config.types.BeaconState.hashTreeRoot(checkpointStateContext.state);
+  }
+  emitter.emit(
+    ChainEvent.checkpoint,
+    {
+      root: config.types.BeaconBlockHeader.hashTreeRoot(blockHeader),
+      epoch: computeEpochAtSlot(config, slot),
+    },
+    checkpointStateContext
+  );
+}
+
+export function emitForkChoiceHeadEvents(
+  emitter: ChainEventEmitter,
+  forkChoice: IForkChoice,
+  head: IBlockSummary,
+  oldHead: IBlockSummary
+): void {
+  const headRoot = head.blockRoot;
+  const oldHeadRoot = oldHead.blockRoot;
+  if (!byteArrayEquals(headRoot, oldHeadRoot)) {
+    // new head
+    if (!forkChoice.isDescendant(oldHeadRoot, headRoot)) {
+      // chain reorg
+      const oldHeadHistory = forkChoice.iterateBlockSummaries(oldHeadRoot);
+      const headHistory = forkChoice.iterateBlockSummaries(headRoot);
+      const firstAncestor = headHistory.find((summary) => oldHeadHistory.includes(summary));
+      const distance = oldHead.slot - (firstAncestor?.slot ?? oldHead.slot);
+      emitter.emit(ChainEvent.forkChoiceReorg, head, oldHead, distance);
+    }
+    emitter.emit(ChainEvent.forkChoiceHead, head);
+  }
+}
+
+export function emitBlockEvent(emitter: ChainEventEmitter, job: IBlockJob, postCtx: ITreeStateContext): void {
+  emitter.emit(ChainEvent.block, job.signedBlock, postCtx, job);
+}
